@@ -9,8 +9,6 @@ var _extends = Object.assign || function (target) { for (var i = 1; i < argument
 
 var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
 
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
 var _formDataToObject = require('form-data-to-object');
 
 var _formDataToObject2 = _interopRequireDefault(_formDataToObject);
@@ -45,6 +43,8 @@ function _possibleConstructorReturn(self, call) { if (!self) { throw new Referen
 
 function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
+/* eslint-disable react/no-unused-state, react/default-props-match-prop-types */
+
 var Formsy = function (_React$Component) {
   _inherits(Formsy, _React$Component);
 
@@ -53,137 +53,104 @@ var Formsy = function (_React$Component) {
 
     var _this = _possibleConstructorReturn(this, (Formsy.__proto__ || Object.getPrototypeOf(Formsy)).call(this, props));
 
-    _this.state = {
-      isValid: true,
-      isSubmitting: false,
-      canChange: false
-    };
-    _this.inputs = [];
-    _this.attachToForm = _this.attachToForm.bind(_this);
-    _this.detachFromForm = _this.detachFromForm.bind(_this);
-    _this.getCurrentValues = _this.getCurrentValues.bind(_this);
-    _this.getPristineValues = _this.getPristineValues.bind(_this);
-    _this.isChanged = _this.isChanged.bind(_this);
-    _this.isFormDisabled = _this.isFormDisabled.bind(_this);
-    _this.reset = _this.reset.bind(_this);
-    _this.resetInternal = _this.resetInternal.bind(_this);
-    _this.runValidation = _this.runValidation.bind(_this);
-    _this.submit = _this.submit.bind(_this);
-    _this.updateInputsWithError = _this.updateInputsWithError.bind(_this);
-    _this.validate = _this.validate.bind(_this);
-    _this.validateForm = _this.validateForm.bind(_this);
-    return _this;
-  }
-
-  _createClass(Formsy, [{
-    key: 'getChildContext',
-    value: function getChildContext() {
-      var _this2 = this;
-
+    _this.getChildContext = function () {
       return {
         formsy: {
-          attachToForm: this.attachToForm,
-          detachFromForm: this.detachFromForm,
-          validate: this.validate,
-          isFormDisabled: this.isFormDisabled,
+          attachToForm: _this.attachToForm,
+          detachFromForm: _this.detachFromForm,
+          validate: _this.validate,
+          isFormDisabled: _this.isFormDisabled,
           isValidValue: function isValidValue(component, value) {
-            return _this2.runValidation(component, value).isValid;
+            return _this.runValidation(component, value).isValid;
           }
         }
       };
-    }
-  }, {
-    key: 'componentDidMount',
-    value: function componentDidMount() {
-      this.validateForm();
-    }
-  }, {
-    key: 'componentWillUpdate',
-    value: function componentWillUpdate() {
+    };
+
+    _this.componentDidMount = function () {
+      _this.validateForm();
+    };
+
+    _this.componentWillUpdate = function () {
       // Keep a reference to input names before form updates,
       // to check if inputs has changed after render
-      this.prevInputNames = this.inputs.map(function (component) {
+      _this.prevInputNames = _this.inputs.map(function (component) {
         return component.props.name;
       });
-    }
-  }, {
-    key: 'componentDidUpdate',
-    value: function componentDidUpdate() {
-      if (this.props.validationErrors && _typeof(this.props.validationErrors) === 'object' && Object.keys(this.props.validationErrors).length > 0) {
-        this.setInputValidationErrors(this.props.validationErrors);
+    };
+
+    _this.componentDidUpdate = function () {
+      if (_this.props.validationErrors && _typeof(_this.props.validationErrors) === 'object' && Object.keys(_this.props.validationErrors).length > 0) {
+        _this.setInputValidationErrors(_this.props.validationErrors);
       }
 
-      var newInputNames = this.inputs.map(function (component) {
+      var newInputNames = _this.inputs.map(function (component) {
         return component.props.name;
       });
-      if (_utils2.default.arraysDiffer(this.prevInputNames, newInputNames)) {
-        this.validateForm();
+      if (_utils2.default.arraysDiffer(_this.prevInputNames, newInputNames)) {
+        _this.validateForm();
       }
-    }
-  }, {
-    key: 'getCurrentValues',
-    value: function getCurrentValues() {
-      return this.inputs.reduce(function (data, component) {
+    };
+
+    _this.getCurrentValues = function () {
+      return _this.inputs.reduce(function (data, component) {
         var name = component.props.name;
+
         var dataCopy = Object.assign({}, data); // avoid param reassignment
         dataCopy[name] = component.state.value;
         return dataCopy;
       }, {});
-    }
-  }, {
-    key: 'getModel',
-    value: function getModel() {
-      var currentValues = this.getCurrentValues();
-      return this.mapModel(currentValues);
-    }
-  }, {
-    key: 'getPristineValues',
-    value: function getPristineValues() {
-      return this.inputs.reduce(function (data, component) {
+    };
+
+    _this.getModel = function () {
+      var currentValues = _this.getCurrentValues();
+      return _this.mapModel(currentValues);
+    };
+
+    _this.getPristineValues = function () {
+      return _this.inputs.reduce(function (data, component) {
         var name = component.props.name;
+
         var dataCopy = Object.assign({}, data); // avoid param reassignment
         dataCopy[name] = component.props.value;
         return dataCopy;
       }, {});
-    }
-  }, {
-    key: 'setFormPristine',
-    value: function setFormPristine(isPristine) {
-      this.setState({
+    };
+
+    _this.setFormPristine = function (isPristine) {
+      _this.setState({
         formSubmitted: !isPristine
       });
 
       // Iterate through each component and set it as pristine
       // or "dirty".
-      this.inputs.forEach(function (component) {
+      _this.inputs.forEach(function (component) {
         component.setState({
           formSubmitted: !isPristine,
           isPristine: isPristine
         });
       });
-    }
-  }, {
-    key: 'setInputValidationErrors',
-    value: function setInputValidationErrors(errors) {
-      this.inputs.forEach(function (component) {
+    };
+
+    _this.setInputValidationErrors = function (errors) {
+      _this.inputs.forEach(function (component) {
         var name = component.props.name;
+
         var args = [{
           isValid: !(name in errors),
           validationError: typeof errors[name] === 'string' ? [errors[name]] : errors[name]
         }];
         component.setState.apply(component, args);
       });
-    }
-  }, {
-    key: 'isFormDisabled',
-    value: function isFormDisabled() {
-      return this.props.disabled;
-    }
-  }, {
-    key: 'mapModel',
-    value: function mapModel(model) {
-      if (this.props.mapping) {
-        return this.props.mapping(model);
+    };
+
+    _this.isFormDisabled = function () {
+      return _this.props.disabled;
+    };
+
+    _this.mapModel = function (model) {
+      if (_this.props.mapping) {
+        return _this.props.mapping(model);
       }
 
       return _formDataToObject2.default.toObj(Object.keys(model).reduce(function (mappedModel, key) {
@@ -196,57 +163,49 @@ var Formsy = function (_React$Component) {
         }
         return mappedModel;
       }, {}));
-    }
-  }, {
-    key: 'reset',
-    value: function reset(data) {
-      this.setFormPristine(true);
-      this.resetModel(data);
-    }
-  }, {
-    key: 'resetInternal',
-    value: function resetInternal(event) {
+    };
+
+    _this.reset = function (data) {
+      _this.setFormPristine(true);
+      _this.resetModel(data);
+    };
+
+    _this.resetInternal = function (event) {
       event.preventDefault();
-      this.reset();
-      if (this.props.onReset) {
-        this.props.onReset();
+      _this.reset();
+      if (_this.props.onReset) {
+        _this.props.onReset();
       }
-    }
+    };
 
-    // Reset each key in the model to the original / initial / specified value
-
-  }, {
-    key: 'resetModel',
-    value: function resetModel(data) {
-      this.inputs.forEach(function (component) {
+    _this.resetModel = function (data) {
+      _this.inputs.forEach(function (component) {
         var name = component.props.name;
+
         if (data && Object.prototype.hasOwnProperty.call(data, name)) {
           component.setValue(data[name]);
         } else {
           component.resetValue();
         }
       });
-      this.validateForm();
-    }
+      _this.validateForm();
+    };
 
-    // Checks validation on current value or a passed value
-
-  }, {
-    key: 'runValidation',
-    value: function runValidation(component) {
-      var _this3 = this;
-
+    _this.runValidation = function (component) {
       var value = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : component.state.value;
 
-      var currentValues = this.getCurrentValues();
-      var validationErrors = component.props.validationErrors;
-      var validationError = component.props.validationError;
+      var currentValues = _this.getCurrentValues();
+      var _component$props = component.props,
+          validationError = _component$props.validationError,
+          validationErrors = _component$props.validationErrors;
+
 
       var validationResults = _utils2.default.runRules(value, currentValues, component.validations, _validationRules2.default);
+
       var requiredResults = _utils2.default.runRules(value, currentValues, component.requiredValidations, _validationRules2.default);
 
       var isRequired = Object.keys(component.requiredValidations).length ? !!requiredResults.success.length : false;
-      var isValid = !validationResults.failed.length && !(this.props.validationErrors && this.props.validationErrors[component.props.name]);
+      var isValid = !validationResults.failed.length && !(_this.props.validationErrors && _this.props.validationErrors[component.props.name]);
 
       return {
         isRequired: isRequired,
@@ -260,8 +219,8 @@ var Formsy = function (_React$Component) {
             return validationResults.errors;
           }
 
-          if (_this3.props.validationErrors && _this3.props.validationErrors[component.props.name]) {
-            return typeof _this3.props.validationErrors[component.props.name] === 'string' ? [_this3.props.validationErrors[component.props.name]] : _this3.props.validationErrors[component.props.name];
+          if (_this.props.validationErrors && _this.props.validationErrors[component.props.name]) {
+            return typeof _this.props.validationErrors[component.props.name] === 'string' ? [_this.props.validationErrors[component.props.name]] : _this.props.validationErrors[component.props.name];
           }
 
           if (isRequired) {
@@ -280,49 +239,31 @@ var Formsy = function (_React$Component) {
           return undefined;
         }()
       };
-    }
+    };
 
-    // Method put on each input component to register
-    // itself to the form
-
-  }, {
-    key: 'attachToForm',
-    value: function attachToForm(component) {
-      if (this.inputs.indexOf(component) === -1) {
-        this.inputs.push(component);
+    _this.attachToForm = function (component) {
+      if (_this.inputs.indexOf(component) === -1) {
+        _this.inputs.push(component);
       }
 
-      this.validate(component);
-    }
+      _this.validate(component);
+    };
 
-    // Method put on each input component to unregister
-    // itself from the form
-
-  }, {
-    key: 'detachFromForm',
-    value: function detachFromForm(component) {
-      var componentPos = this.inputs.indexOf(component);
+    _this.detachFromForm = function (component) {
+      var componentPos = _this.inputs.indexOf(component);
 
       if (componentPos !== -1) {
-        this.inputs = this.inputs.slice(0, componentPos).concat(this.inputs.slice(componentPos + 1));
+        _this.inputs = _this.inputs.slice(0, componentPos).concat(_this.inputs.slice(componentPos + 1));
       }
 
-      this.validateForm();
-    }
+      _this.validateForm();
+    };
 
-    // Checks if the values have changed from their initial value
+    _this.isChanged = function () {
+      return !_utils2.default.isSame(_this.getPristineValues(), _this.getCurrentValues());
+    };
 
-  }, {
-    key: 'isChanged',
-    value: function isChanged() {
-      return !_utils2.default.isSame(this.getPristineValues(), this.getCurrentValues());
-    }
-
-    // Update model, submit to url prop and send the model
-
-  }, {
-    key: 'submit',
-    value: function submit(event) {
+    _this.submit = function (event) {
       if (event && event.preventDefault) {
         event.preventDefault();
       }
@@ -330,53 +271,39 @@ var Formsy = function (_React$Component) {
       // Trigger form as not pristine.
       // If any inputs have not been touched yet this will make them dirty
       // so validation becomes visible (if based on isPristine)
-      this.setFormPristine(false);
-      var model = this.getModel();
-      this.props.onSubmit(model, this.resetModel, this.updateInputsWithError);
-      if (this.state.isValid) {
-        this.props.onValidSubmit(model, this.resetModel, this.updateInputsWithError);
+      _this.setFormPristine(false);
+      var model = _this.getModel();
+      _this.props.onSubmit(model, _this.resetModel, _this.updateInputsWithError);
+      if (_this.state.isValid) {
+        _this.props.onValidSubmit(model, _this.resetModel, _this.updateInputsWithError);
       } else {
-        this.props.onInvalidSubmit(model, this.resetModel, this.updateInputsWithError);
+        _this.props.onInvalidSubmit(model, _this.resetModel, _this.updateInputsWithError);
       }
-    }
+    };
 
-    // Go through errors from server and grab the components
-    // stored in the inputs map. Change their state to invalid
-    // and set the serverError message
-
-  }, {
-    key: 'updateInputsWithError',
-    value: function updateInputsWithError(errors) {
-      var _this4 = this;
-
+    _this.updateInputsWithError = function (errors) {
       Object.keys(errors).forEach(function (name) {
-        var component = _utils2.default.find(_this4.inputs, function (input) {
+        var component = _utils2.default.find(_this.inputs, function (input) {
           return input.props.name === name;
         });
         if (!component) {
           throw new Error('You are trying to update an input that does not exist. Verify errors object with input names. ' + JSON.stringify(errors));
         }
         var args = [{
-          isValid: _this4.props.preventExternalInvalidation,
+          isValid: _this.props.preventExternalInvalidation,
           externalError: typeof errors[name] === 'string' ? [errors[name]] : errors[name]
         }];
         component.setState.apply(component, args);
       });
-    }
+    };
 
-    // Use the binded values and the actual input value to
-    // validate the input and set its state. Then check the
-    // state of the form itself
-
-  }, {
-    key: 'validate',
-    value: function validate(component) {
+    _this.validate = function (component) {
       // Trigger onChange
-      if (this.state.canChange) {
-        this.props.onChange(this.getCurrentValues(), this.isChanged());
+      if (_this.state.canChange) {
+        _this.props.onChange(_this.getCurrentValues(), _this.isChanged());
       }
 
-      var validation = this.runValidation(component);
+      var validation = _this.runValidation(component);
       // Run through the validations, split them up and call
       // the validator IF there is a value or it is required
       component.setState({
@@ -384,44 +311,37 @@ var Formsy = function (_React$Component) {
         isRequired: validation.isRequired,
         validationError: validation.error,
         externalError: null
-      }, this.validateForm);
-    }
+      }, _this.validateForm);
+    };
 
-    // Validate the form by going through all child input components
-    // and check their state
-
-  }, {
-    key: 'validateForm',
-    value: function validateForm() {
-      var _this5 = this;
-
+    _this.validateForm = function () {
       // We need a callback as we are validating all inputs again. This will
       // run when the last component has set its state
       var onValidationComplete = function onValidationComplete() {
-        var allIsValid = _this5.inputs.every(function (component) {
+        var allIsValid = _this.inputs.every(function (component) {
           return component.state.isValid;
         });
 
-        _this5.setState({
+        _this.setState({
           isValid: allIsValid
         });
 
         if (allIsValid) {
-          _this5.props.onValid();
+          _this.props.onValid();
         } else {
-          _this5.props.onInvalid();
+          _this.props.onInvalid();
         }
 
         // Tell the form that it can start to trigger change events
-        _this5.setState({
+        _this.setState({
           canChange: true
         });
       };
 
       // Run validation again in case affected by other inputs. The
       // last component validated will run the onValidationComplete callback
-      this.inputs.forEach(function (component, index) {
-        var validation = _this5.runValidation(component);
+      _this.inputs.forEach(function (component, index) {
+        var validation = _this.runValidation(component);
         if (validation.isValid && component.state.externalError) {
           validation.isValid = false;
         }
@@ -430,54 +350,95 @@ var Formsy = function (_React$Component) {
           isRequired: validation.isRequired,
           validationError: validation.error,
           externalError: !validation.isValid && component.state.externalError ? component.state.externalError : null
-        }, index === _this5.inputs.length - 1 ? onValidationComplete : null);
+        }, index === _this.inputs.length - 1 ? onValidationComplete : null);
       });
 
       // If there are no inputs, set state where form is ready to trigger
       // change event. New inputs might be added later
-      if (!this.inputs.length) {
-        this.setState({
+      if (!_this.inputs.length) {
+        _this.setState({
           canChange: true
         });
       }
-    }
-  }, {
-    key: 'render',
-    value: function render() {
-      var _props = this.props,
-          getErrorMessage = _props.getErrorMessage,
-          getErrorMessages = _props.getErrorMessages,
-          getValue = _props.getValue,
-          hasValue = _props.hasValue,
-          isFormDisabled = _props.isFormDisabled,
-          isFormSubmitted = _props.isFormSubmitted,
-          isPristine = _props.isPristine,
-          isRequired = _props.isRequired,
-          isValid = _props.isValid,
-          isValidValue = _props.isValidValue,
-          mapping = _props.mapping,
-          onChange = _props.onChange,
-          onInvalidSubmit = _props.onInvalidSubmit,
-          onInvalid = _props.onInvalid,
-          onReset = _props.onReset,
-          onSubmit = _props.onSubmit,
-          onValid = _props.onValid,
-          onValidSubmit = _props.onValidSubmit,
-          preventExternalInvalidation = _props.preventExternalInvalidation,
-          resetValue = _props.resetValue,
-          setValidations = _props.setValidations,
-          setValue = _props.setValue,
-          showError = _props.showError,
-          showRequired = _props.showRequired,
-          validationErrors = _props.validationErrors,
-          nonFormsyProps = _objectWithoutProperties(_props, ['getErrorMessage', 'getErrorMessages', 'getValue', 'hasValue', 'isFormDisabled', 'isFormSubmitted', 'isPristine', 'isRequired', 'isValid', 'isValidValue', 'mapping', 'onChange', 'onInvalidSubmit', 'onInvalid', 'onReset', 'onSubmit', 'onValid', 'onValidSubmit', 'preventExternalInvalidation', 'resetValue', 'setValidations', 'setValue', 'showError', 'showRequired', 'validationErrors']);
+    };
+
+    _this.render = function () {
+      var _this$props = _this.props,
+          getErrorMessage = _this$props.getErrorMessage,
+          getErrorMessages = _this$props.getErrorMessages,
+          getValue = _this$props.getValue,
+          hasValue = _this$props.hasValue,
+          isFormDisabled = _this$props.isFormDisabled,
+          isFormSubmitted = _this$props.isFormSubmitted,
+          isPristine = _this$props.isPristine,
+          isRequired = _this$props.isRequired,
+          isValid = _this$props.isValid,
+          isValidValue = _this$props.isValidValue,
+          mapping = _this$props.mapping,
+          onChange = _this$props.onChange,
+          onInvalidSubmit = _this$props.onInvalidSubmit,
+          onInvalid = _this$props.onInvalid,
+          onReset = _this$props.onReset,
+          onSubmit = _this$props.onSubmit,
+          onValid = _this$props.onValid,
+          onValidSubmit = _this$props.onValidSubmit,
+          preventExternalInvalidation = _this$props.preventExternalInvalidation,
+          resetValue = _this$props.resetValue,
+          setValidations = _this$props.setValidations,
+          setValue = _this$props.setValue,
+          showError = _this$props.showError,
+          showRequired = _this$props.showRequired,
+          validationErrors = _this$props.validationErrors,
+          nonFormsyProps = _objectWithoutProperties(_this$props, ['getErrorMessage', 'getErrorMessages', 'getValue', 'hasValue', 'isFormDisabled', 'isFormSubmitted', 'isPristine', 'isRequired', 'isValid', 'isValidValue', 'mapping', 'onChange', 'onInvalidSubmit', 'onInvalid', 'onReset', 'onSubmit', 'onValid', 'onValidSubmit', 'preventExternalInvalidation', 'resetValue', 'setValidations', 'setValue', 'showError', 'showRequired', 'validationErrors']);
 
       return _react2.default.createElement('form', _extends({
-        onReset: this.resetInternal,
-        onSubmit: this.submit
-      }, nonFormsyProps), this.props.children);
-    }
-  }]);
+        onReset: _this.resetInternal,
+        onSubmit: _this.submit
+      }, nonFormsyProps), _this.props.children);
+    };
+
+    _this.state = {
+      isValid: true,
+      isSubmitting: false,
+      canChange: false
+    };
+    _this.inputs = [];
+    return _this;
+  }
+
+  // Reset each key in the model to the original / initial / specified value
+
+
+  // Checks validation on current value or a passed value
+
+
+  // Method put on each input component to register
+  // itself to the form
+
+
+  // Method put on each input component to unregister
+  // itself from the form
+
+
+  // Checks if the values have changed from their initial value
+
+
+  // Update model, submit to url prop and send the model
+
+
+  // Go through errors from server and grab the components
+  // stored in the inputs map. Change their state to invalid
+  // and set the serverError message
+
+
+  // Use the binded values and the actual input value to
+  // validate the input and set its state. Then check the
+  // state of the form itself
+
+
+  // Validate the form by going through all child input components
+  // and check their state
+
 
   return Formsy;
 }(_react2.default.Component);
@@ -529,7 +490,6 @@ Formsy.propTypes = {
   isValid: _propTypes2.default.func,
   isValidValue: _propTypes2.default.func,
   mapping: _propTypes2.default.func,
-  preventExternalInvalidation: _propTypes2.default.bool,
   onChange: _propTypes2.default.func,
   onInvalid: _propTypes2.default.func,
   onInvalidSubmit: _propTypes2.default.func,
@@ -537,6 +497,7 @@ Formsy.propTypes = {
   onSubmit: _propTypes2.default.func,
   onValid: _propTypes2.default.func,
   onValidSubmit: _propTypes2.default.func,
+  preventExternalInvalidation: _propTypes2.default.bool,
   resetValue: _propTypes2.default.func,
   setValidations: _propTypes2.default.func,
   setValue: _propTypes2.default.func,
@@ -555,9 +516,14 @@ var addValidationRule = function addValidationRule(name, func) {
 
 var withFormsy = _Wrapper2.default;
 
+var didWarnAboutWrapperDeprecation = false;
+
 var deprecatedWrapper = function deprecatedWrapper(Component) {
-  // eslint-disable-next-line no-console
-  console.warn('Wrapper has been renamed to withFormsy. Importing Wrapper from formsy-react is depreacted and will be removed in the future. Please rename your Wrapper imports to withFormsy.');
+  if (!didWarnAboutWrapperDeprecation) {
+    // eslint-disable-next-line no-console
+    console.warn('Wrapper has been renamed to withFormsy. Importing Wrapper from formsy-react is depreacted and will be removed in the future. Please rename your Wrapper imports to withFormsy.');
+    didWarnAboutWrapperDeprecation = true;
+  }
 
   return withFormsy(Component);
 };
